+++ include/texture.h
@@ -18,7 +18,7 @@ uint32_t texture2_sample(texture2 *texture, vec2 uv);
 
 texture2* texture2_load(const char *path);
 
-texture2_binding texture2_binding_create(vec3 *triangle);
+texture2_binding texture2_binding_create(dvec3 *triangle);
 void texture2_binding_destroy(texture2_binding binding);
 void texture2f_binding_free(texture2f_binding binding);
 
@@ -31,15 +31,15 @@ void texture2f_destroy(texture2f *texture);
 void texture2f_reset(texture2f *texture);
 float texture2f_get_max_lumel(texture2f *texture);
 void texture_update_lightmap(texture2f *src, texture2 *dst, float max_lumel);
-void texture2f_write_color_bar(texture2f *texture, vec2 *uv, vec3 bar,
-vec3 color);
+void texture2f_write_color_bar(texture2f *texture, vec2 *uv, dvec3 bar,
+dvec3 color);
 
 vec4 texture2f_sample(texture2f *texture, vec2 uv);
 
 void texture2f_refresh_gpu(texture2f *texture);
 void texture2f_refresh_cpu(texture2f *texture);
 
-texture2f_binding texture2f_binding_create(vec3 *triangle);
+texture2f_binding texture2f_binding_create(dvec3 *triangle);
 void texture2f_binding_destroy(texture2f_binding binding);
 
 void gl_set_texture_parameters(void);
