+++ src/world/world_chunk_border.c
@@ -14,8 +14,8 @@ chunk_border_t chunk_border_fetch(ssize2 pos)
 
     for (size_t i = 0; i < CHUNK_GEN_ITER; i++)
         for (size_t j = 0; j < 2; j++) {
-            res.hor[i][j].vec3 = NULL;
-            res.ver[i][j].vec3 = NULL;
+            res.hor[i][j].dvec3 = NULL;
+            res.ver[i][j].dvec3 = NULL;
         }
     adj = world_chunk2d_get(_demo, ssize2_add(pos, (ssize2){-1, 0}));
     if ((adj != NULL) && (*adj != NULL))
